Resolved layout conflicts and added initializers to Food model.
