+++ iOS/Labs8-MealHelper /Labs8-MealHelper/MealEntry/MealSetupTableViewController.swift
@@ -12,7 +12,7 @@ class MealSetupTableViewController: UITableViewController {
     
     // MARK: - Public properties
     
-    var meals = ["Kraft Cheese", "Sausage"]
+    var recipes = ["Kraft Cheese", "Sausage"]
     
     // MARK: - Private properties
 
@@ -43,7 +43,7 @@ class MealSetupTableViewController: UITableViewController {
     }
 
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return meals.count
+        return recipes.count
     }
 
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
@@ -58,6 +58,8 @@ class MealSetupTableViewController: UITableViewController {
     
     @objc func save() {
         print("saved")
+        let meal = Meal(mealTime: "", experience: "", date: "", userId: 1)
+        // API call to /users/:userid/meals
     }
     
     // MARK: - Configuration
@@ -69,19 +71,166 @@ class MealSetupTableViewController: UITableViewController {
         tableView.allowsSelection = false
     }
 
-    
 }
 
 // MARK: - MealSetupTableViewCellDelegate
 
 extension MealSetupTableViewController: MealSetupTableViewCellDelegate {
     
-    func setServingQty(with qty: String, for meal: Any) {
+    func setServingQty(with qty: String, for recipe: Any) {
         // Save qty for meal
     }
     
-    func setServingType(with type: String, for meal: Any) {
+    func setServingType(with type: String, for recipe: Any) {
         // Save type for meal
+    }
+    
+}
+
+class MealSetupHeaderView: UIView {
+    
+    // MARK: - Public properties
+    // TODO: Change to nutrient model
+    var nutrients = [("216 cal", "Calories"), ("30 g", "Carbs"), ("7 g", "Fat"), ("8 g", "Protein")]
+    
+    // MARK: - Private properties
+    
+    private let mainStackView: UIStackView = {
+        let stackView = UIStackView()
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        stackView.axis = .horizontal
+        stackView.distribution = .fillEqually
+        stackView.spacing = 10.0
+        return stackView
+    }()
+    
+    // MARK: - Init
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setupViews()
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    // MARK: - Private methods
+    
+    private func setupViews() {
+        
+        self.addSubview(mainStackView)
+        
+        for nutrient in nutrients {
+            let label = createLabel(with: nutrient.0, subtitle: nutrient.1)
+            mainStackView.addArrangedSubview(label)
+        }
+        
+        mainStackView.anchor(top: self.layoutMarginsGuide.topAnchor, leading: self.layoutMarginsGuide.leadingAnchor, bottom: self.layoutMarginsGuide.bottomAnchor, trailing: self.layoutMarginsGuide.trailingAnchor)
+        
+    }
+    
+    private func createLabel(with text: String, subtitle: String) -> UILabel {
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.numberOfLines = 2
+        label.textAlignment = .center
+        label.text = "\(text)\n\(subtitle)"
+        label.font = UIFont.systemFont(ofSize: 14.0)
+        return label
+    }
+    
+}
+
+class MealSetupFooterView: UIView {
+    
+    // MARK: - Public properties
+    
+    var date: Date?
+    var note: String?
+    
+    // MARK: - Private properties
+    
+    private let mainStackView: UIStackView = {
+        let stackView = UIStackView()
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        stackView.axis = .vertical
+        stackView.distribution = .fill
+        stackView.spacing = 10.0
+        return stackView
+    }()
+    
+    private let notesTextView: UITextView = {
+        let tv = UITextView()
+        tv.translatesAutoresizingMaskIntoConstraints = false
+        tv.font = UIFont.systemFont(ofSize: 17.0)
+        tv.layer.cornerRadius = 8.0
+        tv.layer.masksToBounds = true
+        tv.text = "Add a note..."
+        tv.textColor = UIColor.lightGray
+        return tv
+    }()
+    
+    private lazy var dateTextField: PickerInputField = {
+        let inputField = PickerInputField(defaultValue: Utils().dateAndTimeString(for: Date()))
+        let datePicker = UIDatePicker()
+        datePicker.datePickerMode = .dateAndTime
+        datePicker.addTarget(self, action: #selector(setDateTextField), for: .allEvents)
+        inputField.inputView = datePicker
+        return inputField
+    }()
+    
+    // MARK: - Init
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setupViews()
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    // MARK: - Actions
+    
+    @objc private func setDateTextField(_ sender: UIDatePicker) {
+        dateTextField.text = Utils().dateAndTimeString(for: sender.date)
+    }
+    
+    // MARK: - Private methods
+    
+    private func setupViews() {
+        
+        // Main view
+        self.addSubview(mainStackView)
+        mainStackView.addArrangedSubview(notesTextView)
+        mainStackView.addArrangedSubview(dateTextField)
+        
+        mainStackView.anchor(top: self.layoutMarginsGuide.topAnchor, leading: self.layoutMarginsGuide.leadingAnchor, bottom: self.layoutMarginsGuide.bottomAnchor, trailing: self.layoutMarginsGuide.trailingAnchor, padding: UIEdgeInsets(top: 16.0, left: 0.0, bottom: 0.0, right: 0.0))
+        
+        self.backgroundColor = UIColor.lightGray
+        
+        notesTextView.delegate = self
+    }
+    
+}
+
+// MARK: - UITextViewDelegate
+
+extension MealSetupFooterView: UITextViewDelegate {
+    
+    func textViewDidBeginEditing(_ textView: UITextView) {
+        if textView.textColor == UIColor.lightGray {
+            textView.text = nil
+            textView.textColor = .black
+        }
+    }
+    
+    func textViewDidEndEditing(_ textView: UITextView) {
+        if textView.text.isEmpty {
+            textView.text = "Add a note..."
+            textView.textColor = UIColor.lightGray
+        }
     }
     
 }
